4.21

+++ osa4/controllers/blogs.js
@@ -19,14 +19,11 @@ blogRouter.get('/', async (request, response) => {
 blogRouter.post('/', async (request, response) => {
   const body = request.body
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  console.log(decodedToken)
   if (!decodedToken || !decodedToken.id) {
     return response.status(401).json({
       error: 'token missing or invalid'
     })
   }
-
-
   const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
@@ -47,8 +44,30 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
+
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  console.log(user)
+
+  if (user._id.toString() === blog.user.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({
+      error: 'invalid user'
+    })
+  }
+
 })
 
 blogRouter.put('/:id', async (request, response) => {
